fix: refactorización del modalEmpleados, eliminar logs para depuración y quitar el uso del prop deprecado "item".
