+++ src/hooks/Empleados/useAccionesEmpleado.js
@@ -1,87 +1,169 @@
+//RF19 - Actualizar empleado - https://codeandco-wiki.netlify.app/docs/proyectos/textiles/documentacion/requisitos/RF19
 import { useState } from 'react';
+import dayjs from 'dayjs';
 import { RepositorioActualizarEmpleado } from '@Repositorios/Empleados/RepositorioActualizarEmpleado';
 import { validarDatosActualizarEmpleado } from '@Modelos/Empleados/ActualizarEmpleado';
 
-export const useAccionesEmpleado = () => {
-  const [errores, setErrores] = useState({});
+export const useAccionesEmpleado = (empleadoInicial = null) => {
+  const esEdicion = !!empleadoInicial;
+  const [erroresValidacion, setErroresValidacion] = useState({});
   const [cargando, setCargando] = useState(false);
+  const [alerta, setAlerta] = useState(null);
+
+  // Inicializar con datos de edición o vacío
+  const [datosEmpleado, setDatosEmpleado] = useState(() => {
+    if (esEdicion) {
+      let fechaAntiguedad = null;
+      if (empleadoInicial.antiguedad) {
+        const fecha = dayjs(empleadoInicial.antiguedad);
+        fechaAntiguedad = fecha.isValid() ? fecha : null;
+      }
+
+      return {
+        ...empleadoInicial,
+        idEmpleado: empleadoInicial.id,
+        nombreCompleto: empleadoInicial.nombreCompleto,
+        correoElectronico: empleadoInicial.correoElectronico,
+        antiguedad: fechaAntiguedad,
+      };
+    }
+
+    return {
+      nombre: '',
+      nombreCompleto: '',
+      correoElectronico: '',
+      idEmpleado: '',
+      idUsuario: '',
+      numeroEmergencia: '',
+      areaTrabajo: '',
+      posicion: '',
+      cantidadPuntos: '',
+      antiguedad: null,
+    };
+  });
 
   const CAMPO_OBLIGATORIO = 'Este campo es obligatorio';
 
-  const handleGuardar = async (datos, empleadosExistentes = []) => {
-    // Usar el validador universal
-    const nuevosErrores = validarDatosActualizarEmpleado(datos, empleadosExistentes);
-    setErrores(nuevosErrores);
+  // Métodos para manejar cambios
+  const manejarCambio = (evento) => {
+    const { name, value } = evento.target;
 
-    try {
-      // Determinar si es edición o creación basado en la existencia de ID
-      const esEdicion = !!datos.id;
-
-      if (esEdicion) {
-        console.log('Enviando datos para actualización desde hook:', datos);
-
-        await RepositorioActualizarEmpleado.actualizar(datos);
-        return {
-          exito: true,
-          mensaje: 'Empleado actualizado correctamente',
-        };
-      } else {
-        // Lógica para crear
-        console.log('Crear empleado');
-      }
-    } catch (error) {
-      console.error('Error completo:', error);
-      return {
-        exito: false,
-        mensaje: error.message || 'Error al guardar el empleado',
-      };
+    if (name === 'numeroEmergencia') {
+      const soloNumeros = value.replace(/\D/g, '').slice(0, 10);
+      setDatosEmpleado((prev) => ({ ...prev, [name]: soloNumeros }));
+    } else {
+      setDatosEmpleado((prev) => ({ ...prev, [name]: value }));
     }
+  };
 
+  const manejarAntiguedad = (nuevaFecha) => {
+    if (nuevaFecha === null || (nuevaFecha && nuevaFecha.isValid())) {
+      setDatosEmpleado((prev) => ({
+        ...prev,
+        antiguedad: nuevaFecha,
+      }));
+    }
+  };
+
+  const obtenerHelperText = (campo) => {
+    const err = erroresValidacion[campo];
+    if (err) {
+      return typeof err === 'string' ? err : CAMPO_OBLIGATORIO;
+    }
+    if (
+      !esEdicion &&
+      [
+        'idEmpleado',
+        'idUsuario',
+        'numeroEmergencia',
+        'areaTrabajo',
+        'posicion',
+        'cantidadPuntos',
+        'antiguedad',
+      ].includes(campo)
+    ) {
+      return CAMPO_OBLIGATORIO;
+    }
+    return '';
+  };
+
+  const handleGuardar = async () => {
+    const datosProcesados = {
+      ...datosEmpleado,
+      id: esEdicion ? empleadoInicial.id : datosEmpleado.idEmpleado,
+      nombreCompleto: datosEmpleado.nombreCompleto || datosEmpleado.nombre,
+      correoElectronico: datosEmpleado.correoElectronico,
+      antiguedad:
+        datosEmpleado.antiguedad && datosEmpleado.antiguedad.isValid()
+          ? datosEmpleado.antiguedad.format('YYYY-MM-DD')
+          : '',
+    };
+
+    const nuevosErrores = validarDatosActualizarEmpleado(datosProcesados);
     if (Object.keys(nuevosErrores).length > 0) {
-      return {
-        exito: false,
-        mensaje: 'Corrige los errores antes de continuar',
-      };
+      setErroresValidacion(nuevosErrores);
+      setAlerta({
+        tipo: 'error',
+        mensaje: 'Corrige los errores en el formulario antes de guardar.',
+      });
+      return { exito: false };
     }
 
+    setErroresValidacion({});
+    setAlerta(null);
     setCargando(true);
 
     try {
-      // Determinar si es edición o creación basado en la existencia de ID
-      const esEdicion = !!datos.id;
-
-      //let resultado;
-      if (esEdicion) {
-        console.log('Enviando datos para actualización:', datos);
-
-        const datosActualizacion = {
-          ...datos,
-          id: datos.id || datos.idEmpleado, // Asegura que exista un campo 'id'
-        };
-
-        await RepositorioActualizarEmpleado.actualizar(datosActualizacion);
-        return {
-          exito: true,
-          mensaje: 'Empleado actualizado correctamente',
-        };
-      } else {
-        // Lógica para crear
-      }
-    } catch (error) {
-      console.error('Error completo:', error);
+      await RepositorioActualizarEmpleado.actualizar(datosProcesados);
+      setAlerta({
+        tipo: 'success',
+        mensaje: esEdicion ? 'Empleado actualizado correctamente' : 'Empleado creado correctamente',
+      });
       return {
-        exito: false,
-        mensaje: error.message || 'Error al guardar el empleado',
+        exito: true,
+        mensaje: esEdicion ? 'Empleado actualizado correctamente' : 'Empleado creado correctamente',
       };
+    } catch (error) {
+      setAlerta({
+        tipo: 'error',
+        mensaje: error.message || 'Error al guardar el empleado',
+      });
+      return { exito: false, mensaje: error.message || 'Error al guardar el empleado' };
     } finally {
       setCargando(false);
     }
   };
 
+  const limpiarFormulario = () => {
+    setDatosEmpleado({
+      nombre: '',
+      nombreCompleto: '',
+      correoElectronico: '',
+      idEmpleado: '',
+      idUsuario: '',
+      numeroEmergencia: '',
+      areaTrabajo: '',
+      posicion: '',
+      cantidadPuntos: '',
+      antiguedad: null,
+    });
+    setErroresValidacion({});
+    setAlerta(null);
+  };
+
   return {
-    errores,
+    datosEmpleado,
+    setDatosEmpleado,
+    erroresValidacion,
+    alerta,
+    setAlerta,
     cargando,
+    esEdicion,
+    manejarCambio,
+    manejarAntiguedad,
+    obtenerHelperText,
     handleGuardar,
+    limpiarFormulario,
     CAMPO_OBLIGATORIO,
   };
 };
